Fix Status

+++ WrathCombo/CustomCombo/Functions/Status.cs
@@ -18,7 +18,7 @@ namespace WrathCombo.CustomComboNS.Functions
         /// <returns> A value indicating if the effect exists. </returns>
         public static bool HasEffect(ushort effectID) => FindEffect(effectID) is not null;
 
-        public static byte GetBuffStacks(ushort effectId)
+        public static ushort GetBuffStacks(ushort effectId)
         {
             Status? eff = FindEffect(effectId);
             return eff?.Param ?? 0;
